refactor: move story feed components to dapp
