+++ packages/components/src/StoryFeed/index.ts
@@ -1,7 +1,1 @@
-export * from "./Story";
-export * from "./StoryDetails";
-export * from "./StoryFeedItem";
-export * from "./StoryModal";
-export * from "./StoryNewsroomDetails";
 export * from "./StoryNewsroomStatus";
-export * from "./StoryRegistryDetails";
